+++ src/main/java/com/finfamily/configuration/PasswordEncoderConfig.java
@@ -1,0 +1,27 @@
+package com.finfamily.configuration;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+@Configuration
+public class PasswordEncoderConfig {
+
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+
+        return new PasswordEncoder() {
+
+            @Override
+            public String encode(CharSequence rawPassword) {
+
+                return rawPassword.toString();
+            }
+            @Override
+            public boolean matches(CharSequence rawPassword, String encodedPassword) {
+
+                return rawPassword.toString().equals(encodedPassword);
+            }
+        };
+    }
+}
