+++ src/main/java/com/finfamily/configuration/WebConfig.java
@@ -1,0 +1,16 @@
+package com.finfamily.configuration;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+@EnableWebMvc
+public class WebConfig implements WebMvcConfigurer{
+
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        registry.addMapping("/**").allowedMethods("GET","POST","PUT","PUT","DELETE","OPTIONS","HEAD","TRACE","CONNECT");
+    }
+}
