+++ src/main/java/com/finfamily/domain/RangeValues.java
@@ -1,0 +1,16 @@
+package com.finfamily.domain;
+
+public class RangeValues {
+
+    private double firstValue;
+    private double secondValue;
+
+    public double getFirstValue() {
+        return firstValue;
+    }
+
+    public double getSecondValue() {
+        return secondValue;
+    }
+
+}
