+++ src/main/java/com/finfamily/security/PasswordEncoder.java
@@ -1,0 +1,9 @@
+package com.finfamily.security;
+
+public interface PasswordEncoder {
+
+    String encode(CharSequence rawPassword);
+
+    boolean matches(CharSequence rawPassword, String encodedPassword);
+
+}
