--- src/main/java/com/keycorp/KeycorpApplication.java
@@ -1,16 +1,0 @@
-package com.keycorp;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
-
-
-@SpringBootApplication
-public class KeycorpApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(KeycorpApplication.class, args);
-    }
-
-}
