--- src/main/java/com/keycorp/configuration/PasswordEncoderConfig.java
@@ -1,27 +1,0 @@
-package com.keycorp.configuration;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-@Configuration
-public class PasswordEncoderConfig {
-
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-
-        return new PasswordEncoder() {
-
-            @Override
-            public String encode(CharSequence rawPassword) {
-
-                return rawPassword.toString();
-            }
-            @Override
-            public boolean matches(CharSequence rawPassword, String encodedPassword) {
-
-                return rawPassword.toString().equals(encodedPassword);
-            }
-        };
-    }
-}
