--- src/main/java/com/keycorp/domain/Password.java
@@ -1,34 +1,0 @@
-package com.keycorp.domain;
-
-
-public class Password {
-    private int idUsuario;
-
-    private String basePassword;
-
-    private String newPassword;
-
-    public int getIdUsuario() {
-        return idUsuario;
-    }
-
-    public void setIdUsuario(int idUsuario) {
-        this.idUsuario = idUsuario;
-    }
-
-    public String getBasePassword() {
-        return basePassword;
-    }
-
-    public void setBasePassword(String basePassword) {
-        this.basePassword = basePassword;
-    }
-
-    public String getNewPassword() {
-        return newPassword;
-    }
-
-    public void setNewPassword(String newPassword) {
-        this.newPassword = newPassword;
-    }
-}
