--- src/main/java/com/keycorp/domain/RangeValues.java
@@ -1,16 +1,0 @@
-package com.keycorp.domain;
-
-public class RangeValues {
-
-    private double firstValue;
-    private double secondValue;
-
-    public double getFirstValue() {
-        return firstValue;
-    }
-
-    public double getSecondValue() {
-        return secondValue;
-    }
-
-}
