--- src/main/java/com/keycorp/domain/Usuario.java
@@ -1,109 +1,0 @@
-package com.keycorp.domain;
-
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.client.RestTemplate;
-
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-import java.util.List;
-
-
-@Entity
-@Embeddable
-public class Usuario {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-
-    private int idUsuario;
-
-
-    private String nome;
-
-
-    private String cpf;
-
-
-    private String dataNascimento;
-
-
-    private String email;
-
-
-    private String senha;
-
-
-    private LocalDateTime dataAtualizacao;
-
-    public Usuario(int idUsuario, String nome, String cpf, String dataNascimento, String email, String senha, LocalDateTime dataAtualizacao) {
-        this.idUsuario = idUsuario;
-        this.nome = nome;
-        this.cpf = cpf;
-        this.dataNascimento = dataNascimento;
-        this.email = email;
-        this.senha = senha;
-        this.dataAtualizacao = dataAtualizacao;
-    }
-
-    public int getIdUsuario() {
-        return idUsuario;
-    }
-
-    public void setIdUsuario(int idUsuario) {
-        this.idUsuario = idUsuario;
-    }
-
-    public String getNome() {
-        return nome;
-    }
-
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-
-    public String getCpf() {
-        return cpf;
-    }
-
-    public void setCpf(String cpf) {
-        this.cpf = cpf;
-    }
-
-    public String getDataNascimento() {
-        return dataNascimento;
-    }
-
-    public void setDataNascimento(String dataNascimento) {
-        this.dataNascimento = dataNascimento;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getSenha() {
-        return senha;
-    }
-
-    public void setSenha(String senha) {
-        this.senha = senha;
-    }
-
-    public LocalDateTime getDataAtualizacao() {
-        return dataAtualizacao;
-    }
-
-    public void setDataAtualizacao(LocalDateTime dataAtualizacao) {
-        this.dataAtualizacao = dataAtualizacao;
-    }
-
-    public Usuario(){}
-
-}
