--- src/main/java/com/keycorp/security/PasswordEncoder.java
@@ -1,9 +1,0 @@
-package com.keycorp.security;
-
-public interface PasswordEncoder {
-
-    String encode(CharSequence rawPassword);
-
-    boolean matches(CharSequence rawPassword, String encodedPassword);
-
-}
