+++ src/test/java/com/finfamily/FinFamilySendEmailTests.java
@@ -1,0 +1,18 @@
+package com.finfamily;
+
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+
+public class FinFamilySendEmailTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
