+++ src/test/java/com/finfamily/domain/TodosUsuariosTests.java
@@ -1,0 +1,16 @@
+package com.finfamily.domain;
+
+import com.finfamily.controller.CredenciaisTests;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
+import org.springframework.stereotype.Repository;
+
+
+@Repository
+public interface TodosUsuariosTests extends JpaRepository<UsuarioTests, Long > {
+    //linguagem HQL ou JPQL para escrever os scripts do banco
+
+    @Query("select u from Usuario u where u.credenciais = :credenciais")
+    UsuarioTests buscarUsando(@Param("credenciais") CredenciaisTests credenciaisTests);
+}
