--- src/test/java/com/keycorp/KeycorpSendEmailTests.java
@@ -1,18 +1,0 @@
-package com.keycorp;
-
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-
-public class KeycorpSendEmailTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
